Update to use two separate dictionary files
One dictionary (popular.txt) is used to select a random secret word,
while the other dictionary (enable1.txt) is used to validate guesses.

Note that it's critical, but not validated, that the secret words
dictionary is a strict subset of the validation dictionary.  Otherwise,
the correct secret word may not be a valid guess.